+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - 📖 documentation on the integration tests
 
+## Changed
+- ♻️ refactored poetry build logic
+
 ## Fixed
 - 📖 indents in quickstart doc
 - 📝 add integration tests to the quickstart structure